requirements.txt
grpcio
grpcio-tools
opencv-python-headless
numpy
protobuf
redis
transformers
torch
torchvision
Pillow
inference-sdk
